Code_organization
1.Code organization
2.Creating an output folder if it does not exist
